Move index.js into src folder
