+++ spec/init.spec.js
@@ -1,5 +1,5 @@
 describe('FormSG SDK', () => {
   it('should be able to initialise without arguments', () => {
-    expect(() => require('../index')()).not.toThrow()
+    expect(() => require('../src/index')()).not.toThrow()
   })
 })
